updated ImageController and ImageService Java files for session management

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,31 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
-        Image image = imageService.getImage(imageId);
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
+        //track user session.
+        User user = (User) session.getAttribute("loggeduser");
+        //Image image = imageService.getImage(imageId);
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        String error = "Image owner can only edit the image";
+
+        if (!user.getId().equals(imageService.getUserId(imageId))) {
+            Image image = imageService.getImage(imageId);
+            model.addAttribute("image", image);
+            model.addAttribute("editError", error);
+            return "images/image";
+        }
+        else {
+            Image image = imageService.getImage(imageId);
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+
+        //String tags = convertTagsToString(image.getTags());
+        //model.addAttribute("image", image);
+        //model.addAttribute("tags", tags);
+        //return "images/edit";
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -113,7 +131,7 @@ public class ImageController {
     //The method also receives tags parameter which is a string of all the tags separated by a comma using the annotation @RequestParam
     //The method converts the string to a list of all the tags using findOrCreateTags() method and sets the tags attribute of an image as a list of all the tags
     @RequestMapping(value = "/editImage", method = RequestMethod.PUT)
-    public String editImageSubmit(@RequestParam("file") MultipartFile file, @RequestParam("imageId") Integer imageId, @RequestParam("tags") String tags, Image updatedImage, HttpSession session) throws IOException {
+    public String editImageSubmit(@RequestParam("file") MultipartFile file, @RequestParam("imageId") Integer imageId, @RequestParam("tags") String tags, Image updatedImage, HttpSession session, Model model) throws IOException {
 
         Image image = imageService.getImage(imageId);
         String updatedImageData = convertUploadedFileToBase64(file);
@@ -132,7 +150,9 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        //return "redirect:/images/" + updatedImage.getTitle();
+        model.addAttribute("images", updatedImage);
+        return "redirect:/images/";
     }
 
 
@@ -140,9 +160,23 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        //imageService.deleteImage(imageId);
+        //return "redirect:/images";
+
+        User user = (User) session.getAttribute("loggeduser");
+        String deleteError = "Only the owner of the image can delete the image";
+
+        if (!user.getId().equals(imageService.getUserId(imageId))) {
+            Image image = imageService.getImage(imageId);
+            model.addAttribute("image", image);
+            model.addAttribute("deleteError", deleteError);
+            return "images/image";
+        }
+        else {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
